+++ stories/examples/Collapsible.js
@@ -1,17 +1,31 @@
 import React from 'react'
 import { CollapsibleTheme } from '../theme'
 import { Collapsible } from 'react-ui-core/src/Collapsible'
-import Checkbox from './Checkbox';
+
+export default class Checkbox extends React.Component {
+  render(){
+  	return (<div className={this.props.className}>
+              <label><input type="checkbox"
+                name="checkbox"
+                value={this.props.label} />
+                {this.props.label}
+              </label>
+            </div>)
+  }
+}
 
 let title = 'Amenities';
-let showableItems = ['Air Conditioning', 'Elevator', 'Laundry Facility', 'Disability Access'];
-let nonshowableItems = ['Dishwasher','Garbage Disposal', 'Island Kitchens','Microwave','Refrigerator', 'Stainless Steel Appliances','Washer/Dryer in Unit',];
+let showableItems = ['Air Conditioning', 'Elevator',
+                    'Laundry Facility', 'Disability Access'];
+let nonshowableItems = ['Dishwasher','Garbage Disposal',
+                      'Island Kitchens','Microwave','Refrigerator',
+                      'Stainless Steel Appliances','Washer/Dryer in Unit',];
 
 const renderItem = (item) => {
   return (
-    <Checkbox 
-      className={CollapsibleTheme.checkboxStyle} 
-      label={item} 
+    <Checkbox
+      className={CollapsibleTheme.checkboxStyle}
+      label={item}
     />
     )
 }
@@ -39,8 +53,8 @@ export const Collapse = (
     showableItems={myShowableItems}
     nonshowableItems={myNonShowableItems}
     theme={CollapsibleTheme}
-    togglemore='see all amenities'
-    toggleless='see less amenities'
+    hidden='see all amenities'
+    visible='see less amenities'
     handleClick={handleChange}
   />
   </div>
